requirements.txt
PyTDC
scikit-learn
numpy
rdkit
tqdm